Fix usage of unix sockets on OSX
The call to getnameinfo() fails on OSX if the sockaddr actually means a
unix socket. Linux, apparently, just fills with "localhost", so do we,
if we detect a unix socket.

Simple test:

$ ncat -lUc 'echo $NCAT_REMOTE_ADDR:$NCAT_LOCAL_ADDR' test &
$ ncat -U test

This currently fails with:
Ncat: getnameinfo failed: Undefined error: 0 QUITTING.

With this fix we'll get "localhost:localhost" just like on Linux.