Fix names of RC2 cipher suites in sslv2.nse
The previous naming for suite 0x040080 was wrong according to [1] and [2]. It
is an export suite but its name did not reflect that.

This also removes the extra "CBC_" in both RC2 cipher suites as it seems more
consistent with other names, the specification and OpenSSL.

As the names came from Wireshark, the same issue was reported to them. [3]

[1] https://github.com/openssl/openssl/blob/OpenSSL_1_0_2-stable/ssl/ssl2.h#L95
[2] http://www-archive.mozilla.org/projects/security/pki/nss/ssl/draft02.html
[3] https://code.wireshark.org/review/#/c/14197/