Use the AppleLocale prefs directly
This is really easier and much more portable (Mac OS 10.5 and later) 
than the other solution proposed. The problem with the actual code happens 
when I set English as my main language with a French localisation and 
keyboard. Sometimes, the AppleLanguages configuration file contains names 
like "English" instead of "en", "EN" or "en_EN". Moreover, there can be 
more than one language defined here. Instead, I suggest with this PR to 
use AppleLocale directly, which is in the "en_EN" format (or whatever 
language) and is set to the current language used by the machine. This 
would allow users to just launch Zenmap and have the software launched in 
the desired language without having to use the terminal (they can still 
change their machine language in the System Preferences or via the 
Terminal, but it would be easier to use).