Warn user when passing invalid IPv6 address
When passing an IPv6 address that is link-local (`fe80::/10`), the interface to be used *must* be specified. This can be done with either `-e <iface>` or by appending `%iface` to the address (e.g. `2001:db8::1%eth0`). If the user does neither, they should be warned that this is invalid.

Previously a connect call would later fail with EINVAL (invalid argument). This code prevents that from happening and instead returns a helpful message.