Update Zenmap to consider the -v0 option
Zenmap does not consider the "-v0" option as NmapOptions.py simply uses the argument and renders multiple copies of -v. In this fix, self["-v"] has been assigned the value of -1 for the case of "-v0" because the default value for self["-v"] (when there is no option "-v" in the command itself) has been set to 0.