Enhance run_tests.py so it can be called from anywhere Fix bug for running tests on OS X
python run_tests.py had to be called from inside test directory,
now be called from anywhere, and not hard-coded anymore.

on Linux the directory created inside build/ name is started with
"lib." but on OS X it is just named "lib". removing the dot in
matching pattern will let the user to run the script on OS X as well.