move manual defines away from libdnet's config.h.in
config.h.in is generated by the autotools (and thus overwritten when
using `autoreconf -fi`).  In that process, the defines added to
libdnet-stripped/include/config.h.in in
https://github.com/nmap/nmap/commit/4aa4a154f9b2d6bf143d31d906866456168dcda3#diff-4df48e6040fa37ab449418304adbe1e8
are lost. This causes build failures on some architectures.

This commit moves the defines to route-bsd.c which is the only place
where they are currently needed.