Update / modification to Sping Boot / Spring SVC endpoints
This PR changes all Spring Boot Actuator  and MVC Endpoints  to end NOT with a trailing slash. 

Some work with a trailing slash, some don't -- tested sucessufully with loggers which gives a 401 with "/" but a hole beauty without. (FYI: I am using since a while a script for this and never used a trailing slash).

Morever it adds to Spring Boot a few endpoints which were missing