Fix name resolution for socks4 proxy.
Previously client didn't try to resolve hostname prior to sending
request to proxy. It violates SOCKS4a specification which says
DSTIP should be set to .X if client cannot resolve the destination
host and may be problematic in some environments.
This behaviour can be overriden with --nodns option (to force client to
use socks4a extention);

Bug-Url: https://bugzilla.redhat.com/1573148