Remove FIXME comment: newstrlen *is* checked.
I added this a long time ago, and decided to check it. In fact,
newstrlen is used to calculate newstrend, and each section of the
template is checked to ensure it does not go past newstrend, so the
intent is met and the length is not exceeded. I still think it could be
written more clearly, but it's good for now. #1834