python3: drop use of cmp by using rich comparison methods e.g. __lt__
It's available since python 2.1 and cmp is considered inferior to use in
python2 and totally removed in python3.

For Gtk ListStore sort functions that need a cmp interface, implement cmp
manually as (a > b) - (a < b)
Note: get_addrs must always return a list, since on python3 you cannot
compare a Nonetype to a list. An empty list will always compare < a list
with something in it, which was the python2 behavior when comparing to
None.