python3: port from c?StringIO.StringIO to io.StringIO
The io module was introduced in python 2.6 and in python 3.x fully replaces the
cStringIO/StringIO modules from python 2.x, which are no longer
available.

It is mostly a drop-in replacement, with the caveat that io.StringIO and
io.BytesIO are not interchangeable due to python3 separating the str and
bytes types, but the nmap usage is restricted to manipulating files as
text, so it should work as-is.