Create searchnse
This is a quick script to search nmap scriptdb resident on kali (/usr/share/nmap/scripts/script.db).  I wanted a search simliar to searchsploit but I realized I didn't need it to be as robust.  I created searchnse to quickly parse for a script based on an initial seed term and a possible 2nd filter term.