python3: upgrade most code to import all __future__s
absolute_import and print_function are kind of no-brainers

division seems safe as there don't seem to be many violations of such,
and if there are, it is broken and needs to be fixed. Some errors have
been caught by enforcing this on python2 as well.

unicode_literals is usually safe, but some things were excluded due to
restrictions on needing str() itself. Specifically, pygtk is very picky.
Rather than manually casting many strings back to str(), these files are
simply excluded for now.