fix broken escaping of word
We want to print

assuming "unknown".

like we do in other parts of the file too. But for some reason, it was

assuming \unknown\.

Which is strange, and also fails to compile on python3, or python2 with
unicode_literals, because \unknown isn't a valid unicode \uXXXX escape.

This was broken in commit da0c94700402840bf3ea04d3362699c52f3663d8 which
changed u"assuming \"unknown\"." to u'assuming \unknown\.' with single
quotes, but dropped the quotes instead of the escapes.