Truncate rlim_cur/rlim_max if greater than INT_MAX
On macOS, the `rlim_t` type used by `rlimit` is 64 bits wide.
For me, `getrlimit` actually returns `LONG_MAX` as `rlim_max`,
which makes `maximize_fdlimit` return -1 due to integer
truncation.

For the purposes of the `maximize_fdlimit` function, it should be
sufficient to just truncate the value and keep the return type of
`int` because no one needs that many files in practice.