http-shellshock: Fix false positive by matching at start of line
Running the http-shellshock script on a Zulip installation results in
a false positive because Zulip (intentionally and safely) reflects the
HTTP User-Agent header as the data-platform attribute on a <div> so
that it can be matched with CSS rules.

https://github.com/zulip/zulip/blob/3.0/templates/zerver/portico.html#L11

Change the payload to write the random string with multiple echo
statements, so that the random string won’t match a reflected payload
that was not actually executed.

Signed-off-by: Anders Kaseorg <andersk@mit.edu>