Add a warning per weak cipher in ssl-enum-ciphers
Currently the ssl-enum-ciphers script gives a list of warnings for a given port, which is a concatenation of warnings spotted on the various ciphers available
on that port.
This commit adds a 'weak' or 'insecure' for each cipher encountered.
In xml output, this translates into a 'warning' tag alongside the 'name', 'kex_info' and 'stength' tags.
It aims to give more precise information since some ciphers rated 'A' for cipher strength can still be weak, which is not accounted for currently.