Fix url parse_path() discarding consecutive slashes
The `parse_path` function from `nselib/url.lua` splits a path into
segments using the `/` delimiter. However, it ends up discarding
consecutive `/`s as it doesn't add empty segments to the parsed table.
This patch ensures that we preserve these empty segments and thus are
able to rebuild the correct path with the `build_path` function.