ncat: perform half-duplex shutdown upon EOF
Previously, `netexec()` would cease all communication as soon as it read EOF
from either the child proces *or* the remote side. This meant (for example)
when the other end of the TCP socket calls `shutdown(fd, SHUT_WR)`, the child
process doesn't get a chance to send any more response, drain its buffers, etc.

The new behaviour is to propagate EOF onwards from the child process to the
remote, and vice versa, and keep a flag variable for each side to say whether
we still want to be reading from that side.