Clarified the npcap installation requirements
As it stands, the wording implies that npcap itself needs to be installed for most Nmap scans to work. Given that there are alternative open-source variants available, these nmap scans will still work even if it is not installed, and they are installed instead, so the wording should accurately reflect this.