API version identification for HPE OneView
This is based on the official docs that HPE pointed me at.
I've only got one old server to test it against but it is
fairly simple so I don't see there is much to go wrong.

I pinched the structure for the script from vmware-version so I hope it
is OK.